+++ src/main/java/com/gdu/wacdo/services/CollaboratorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.gdu.wacdo.models.Collaborator;
 
-public interface CollaboratorRepository extends JpaRepository<Collaborator, Integer> {
+public interface CollaboratorRepository extends JpaRepository<Collaborator, Long> {
 
 }
